Use plugin logger instead of Minecraft logger in README example
The JavaPlugin class provides plugins with their own logger via getLogger() which is the preferred means of logging content to the console as it prepends the plugin's name. Using a Logger with the Minecraft prefix is non-descript and poor practice because it will not properly source from where the log message is coming. Beginners should not be given this poor practice as an example.